"Teleporter" key item
Created the "Teleporter" key item which when used teleports you to just outside the last building you healed in. This uses the "Healing Spot" setting of the map in which the player healed-- we'll have to remember to set those in the future.
This involved the creation of a new "NewItems" plugin
Also correctly set the "Home" global metadata property to inside the player's house map