Reworked encouner sets in river route/big cave, traveling trainer now separate trainers
River Route now has a bunch of tall grass, and its encounter set has been split between regular and tall grass
Galarian Ponyta now appears in the river grove instead of Mincinno (which was added in the starting town earlier)
Wild grass added to the southern part of the big cave b1, and to a secretish part of the big cave b2. The encounters set for this area has been completely reworked.
The "traveling trainer" now longer travels. The 2nd encounter with Max at the river route is now a separate encounter with a female ace trainer called Raven.
For both ace trainers, moved their second overworld companion to the side the player approaches them on, so it can be interacted with.
Sing no longer fails if the target is at exactly half health