New move fix and two descriptions, trainer pokemon abilities less random
Added desriptions for Staredown and Inure, and fixed a compile error with Dragon Roar
Ported a system from the old prototype where a trainer of a given name with a pokemon of a given species will always have the same ability by default (if not set explicitly in the trainer.txt). This is to stop the ability from changing randomly between fights against them.