Update pokemon.txt
Wrote in pokedex entries for the mega-turned-regular evolutions, this included:

Mega Medicham
Mega Camurupt
Mega Altaria
Mega Abomasnow
Mega Audina
Gigantamax Sandaconda

as a though - I wonder how sandaconda speaks? its name is such a mouthful that I would think that it would just say "sanda" to not be that annoying to everyone around it.