Style values and editor UI/npc, LevelingOverhaul renamed
Pokemon no longer gain EVs from battle
Style Values are now used in the place of EVs. Each style value is equal to 1 stat in the final stats of a pokemon at the level cap (70).
Pokemon have a default value of 8 in each style value (leaving a default pool of 2 stats). This is also true for trainers.
An NPC in the Boss FIght Test Arena allows you to access a UI which allows you to adjust your style values arbitrarily at no cost. The max of style values total is
Fixed a bug where some pokemon would case the move relearner to crash while checking for relearnable level 1 moves