Compile-time adding of cry behaviour to overworld pokemon, cries replaced and expanded
Added a replacer in the compiler for maps which adds a script to show an overworld pokemon's name and play its cry when you talk to it. This doesn't affect pokemon with "nocry" in their name (used for exceptions like overworld pokemon with the unique behaviour where you can check their held item)
With this change, "overworld" needs to always be the last command in an event's name to compile correctly
Deleted all existing cry sound files, replaced them with ones from v19 of the Gen 8 essentials download pack. This means having cries for pokemons gen 6-8 which were lacking before. Also, everything is in OGG format now rather than WAV.
Fixed a compile error in the AI changes from the last commit around the switch statement