"Seen" status useful again, dexnav improvements, areas page of dex, "available by level" search more advanced
The areas tab of the pokedex now shows the names of the maps the species is an encounter on, as well as any maps for which their pre-evolutions are encounters
Nonlegendary dex entries are now available regardless of having been seen or not. No longer sets all nonlegendary species seen when you grab the dex, so that state can be used for other things.
The dexnav now does not show a silhouette unless a species has been seen at least once.
The dexnav will show that you have seen no pokemon, and not include certain UI elements, when thats the case.
The dexnav now also tells you when you've seen all pokemon on a route, but not yet caught all of them
Removed encounters for the starting town
The search "by available level" now takes into account whats actually encounterable by that point in the game. For example, searching the value 15 in the current prototype shows the pokemon which are available by the first gym (which is the end of level cap 15). The only mistake is that it assumes you have access to all and infinite evolution stones and similar.