Fix to topleft animations bug
AnimationPlayer's, if given no target, will assume the target's sprite origin is in the center of the screen, rather than at the very top left. This is useful in case, despite using a move that has no targets, is using a fallback animation whose cells are set to focus on the target. For example, toxic spikes falling back on the animation of poison powder.