Species name appears on far top right of all dex pages, move categories appear on both moveset pages
Moving the species name to the top right means the "Level up learn set of {species}" titles in the main space can be removed
That said, only removed the ones on the level up and tutor move pages thus far. This gave room to show one more move on each page as well as to show move categories next to the move type.