Avatar compilation overhaul, support for up to three avatar phases
Rather than "Moves" and "PostPrimeMoves" there are "Moves1", "Moves2", and "Moves3" which define the 1st, 2nd, and 3rd phase, respectively.
Compiler now has many more default values for undefined avatar properties, and doesnt bother to write those things out when writing the PBS (default turns of 2, for example)
An error message shows in the console if a defined avatar has an ability it shouldn't be able to get
If a battle would crash due to some sort of runtime error, the message is shown, but then the battle continues with the result being given to the player.
Beast legend avatars reworked