Continuing the move code and AI code refactors
Started the refactor for having the actual damage calcs double as the AI damage calcs by modifying all of the previously isolated damage calculation methods in Move_Usage_Calculations to also work for use in the AI.  For this purpose, Introduced two new methods of the battler class "shouldAbilityApply?" and "shouldTypeApply?" which must take a boolean value to say whether we are asking for the AI or not.
The concept of "trampling minimize" is removed
Deconstructed the pbMoveBaseDamage AI method.  Many of the considerations of that method have been replaced by instances of pbBaseDamageAI defined within the class of a move code. Others have been cut entirely for various reasons
Reorganized some helper methods of the battler class so that they are alongside the other battler class scripts