Almost all 080 to 0FF AI transfered
New generic move class PokeBattle_DrainMove created, all draining moves refactored as subclasses of it. It has pretty snazzy AI
Nervebreak is accounted for in AI code for most healing moves
Made a helper method for PokeBattle_Battler determining if the battler has a move that lets them move while asleep
Made a helper method for PokeBattle_Battle which returns a PokeBattle_Move object for a given move ID. Useful for checking information about moves that only stored there (i.e. ignores burn, is healing move, is protect style move)
Refactored PokeBattle_RecoilMove to have most moves define a recoilFactor method which is then used in pbRecoilDamage, rather than most defining pbRecoilDamage themselves. Also added a getScore to the PokeBattle_RecoilMove class which reduces the score the higher the recoil factor is.
Added a getScore method to the PokeBattle_ProtectMove class with some basic checks
Added helper methods for checking for hazard presence on a particular side. Used to refactor the AI for moves that swap you out or force swaps on an enemy target
Added a helper method for moves that KO the user
Scrubbed the code for two essentials-base move codes that arent used anymore
Added a helper method for determining if a user is capable of using the BUNKERDOWN ability to survive an explosion
Added helper methods to determine if item stealing moves can actually do so