Move_Effects_100_17F scoring transfer progress, many other various refactors tweaks and AI additions
Added battler helping moves for determining if the battler has a substitute present, if they have a sound move, if they have a move which punishes all statuses, and if it is their first turn out in the battle. The substitute one is to make it easier for the AI to not value additional move procs on substituted pokemon
Added scoring methods to some moves that were missing it in the 080_0FF section
Reduced the score of trapping moves by 20 across the board as a sort of "annoying moves tax"
Removed some weird checks that caused status-punishment moves to not double in power against substituted targets. Bulbapedia doesn't mention this behaviour
Re-did the protect move AI because it checked for aspects of the "target" even though the move is self-targeted
The getOpposing method of the battler class now has an optional parameter to restrict the callbacks to enemies that are near the battler
Added a battler helper method which yields every pokemon that could attack the user the next turn. This means battlers that are opposing, near, and aren't recovering from a recharge move. It also has an optional parameter to only return enemies that are capable of attacking you with a phys/spec/status move
Refactored the targeted healing move to be a general healing move method where the target can be the self
Made a parent class for all terrain setting moves, with new AI
Added a move method hasKOEffect? which means it values targeting enemies it thinks it can kill even higher than normal
Added a scoring function for the PokeBattle_TwoTurnMove parent class
Refactored type-applying spike moves as subclasses of a shared parent class. The parent class now also supports the behaviour of attacking moves that set these spikes, and so Cruel Whipblade and Snowy Sorrows have been reworked thereby