Trainer entry inheritance improvements
"Extends" is now used for a longer syntax whereby you specify a type, name, and version of the entry to inherit from
"ExtendsVersion" is used for the previous short syntax where the type and name have to be the same
Can overwrite aspects of individual pokemon on the child entry. Replaces on a per attribute basis, e.g. muse replace the entire moveset at once.
Use "RemovePokemon" to specify that the child class should not include a particular pokemon from the parent clas
For both of the above, two pokemon are assumed to be the same across entries if they have the same species and at least one of the same level or nickname
If a newly defined pokemon in the child class needs to appear on the team before pokemon from the parent class, you can use the "position" attribute to specify this