Effects System Refactor
Finally pulled in the entirety of the battle system code from base essentials into our Battle plugin
Refactored the Item plugin into the Battle plugin
Refactored the structure of the Battle plugin to match base essentials but with some scripts split up or further categorized into folders as I saw appropriate to reduce bloat
Battle effects (on battlers, positions, sides, and the whole field) are no longer tracked as integer indices of an array but as entries in a hash whose key is symbols
Most interactions with the effects system within moves and abilities are now abstracted behind an interface which is created by the EffectHolder module which battlers, positions, sides, and activefield all implement. This makes direct interaction with the effects hash fairly rare
This abstracting layer disallows illegal operations based on the type and location of the effect (for example, not allowed to apply "false" to an integer) and also checks for incorrect effect definitions
Effects are now defined by static calls to a new GameData module subclass BattleEffect, passing in a hash that defines the effect
Every effect in the game has been refactored thusly in the Battle/Effects/Effect Defintions folder
Behaviour such as "what to do when the effect is applied" are now defined in those hashes, to make it so that various moves, abilities, curses, etc. can apply those effects simply and with consistent behaviour
I've made major improvements and reworks to the Auto Battle Testing system, which allowed me to fix hopefully the vast majority of the major errors introduced by this refactor. There may still be many smaller errors such as display bugs, typos, or unintended behaviour
I also finally figured out Rubocop, so this update includes the first instance of Rubocop linted code in the Battler class, I will be experimenting more with finding the layout and style I prefer and will do a rubocop pass over the entire codebase in the near future