Avatar UX experiments
Rewrote the databox code to allow for a mix of sizes on a side at once
Made it so that pokemon with multiple HP bars don't go to a thinner databox in doubles or higher (especially relevant when an avatar summons)
The number of healthbars a pokemon has is no longer based on whether or not its species is a legend, but rather a property you can define in avatars.txt. If it isn't, the number of healthbars is assumed to be the same as the number of phases (which is the number of defined movesets)
Avatars with only one healthbar can only incur one status. For now this only includes Combee and Phione (decided that this could be a cool flavor point? dunno)
Also gave the avatar of gourgeist the 2nd phase I forgot to give it