Trainer switch AI refactor for cleanliness, more debug lines, and more
Trainers more likely to switch when seeded and/or drowsy
New policy SACS_NOT_SWAPS is less likely to switch when in red bar health (25% or less) and also in that situation ignores concerns about switching based on effects that would cause death (e.g. seeded). It doesn't ignore effects that would lead to tempo gain for the player (e.g. drowsy)
Rewrote the code that affected likelihood to switch based on the scoring of ones moves. Rather than merely being very likely to switch if it has bad moves, it not is more or less likely to switch based on just how good its best move is. E.g. Its quite unlikely to switch if it thinks it has a great move
Added a bunch of debug lines in the switch evaluation to make it easier to tell why a trainer is switching
Old unneccesary code for wild battler AI is removed