Small buffs to several Pokemon
Blasfemmie could use a four or five step Defense-boosting move now that Acid Armor was cut so it can run double dance sets again. Could also use a signature move or Ability that makes it sing (figuratively).

Grapploct was given Technician which it has in canon because Energetic was doing nothing on it. Technician mainly buffs Mach Punch, Power-Up Punch, Circle Throw, and the newly-added Aqua Jet. Could be replaced, but seemed low-impact but interesting.