Sandstone System Redux: Removed "Algorrhythmia"
Cut one of the Sandstone System Redux side areas; the themes of "Algorrhythmia" will now be integrated into "Terminal Confusion". This was done to correct a planning error in which I accidentally designed the hub with one more side area than I'd originally intended to create.
Reworked "Sandstone Switchbox" (the hub) accordingly.
Tweaked the details of the circuit paths at the end of "Wave Lengths".
Tweaked "Read Only" to keep its circuitry visually and spatially consistent with the rest.