Alloyed Thicket replant; Mirror Tundra Redux changes
Transplanted a couple trees in Alloyed Thicket to correct an avatar placement issue.
Laid out "Mirror Tundra Redux" by copying "Mirror Tundra" and adjusting it to include the statue-based shortcut mechanic.
Removed the encounter grass in "Endless Ice" and "Mirror Crevasse".
Tweaked the exact position of the shortcuts in "Glacial Labyrinth" and "Frozen Lake".
Replaced some leftover placeholders in "Frozen Lake".
Corrected incorrect climbing rock sprites in "Mirror Maze".
Added a "Reveal Glass" placeholder and removed both Landorus and an unnecessary statue from "Sandstone Switchbox".
Added lots of obstacles to "Howling Frost" to reduce how punishing the wind will be if we implement it as currently planned (intermittent blasts that push the player as far as possible in a given direction).