PORYGON-MAV initial update
BST and learnset adjustment. Quarantine should be evo move once it is implemented.
Changed Realist to Sharpness to make it less of a massive tank (and cuz it has two swords it's a crime not to have it).
Side note: Porygon line (and Spritzee) may want a Tutor adjustment to make the Makyan evo branching feel less like a direct upgrade on move options compared to the standard evo line.