AI does not preserve damaged slowed mons
Added system for evaluating the relative speed of a mon at its level bracket, and penalizing switching if it thinks its slow (and thus needs HP to get turns) and does not have HP to spend.
This results in the AI being "foggier" (harder to predict) switches, since it will still preserve a low health mon if it has a very good switchin. (And the AI can break out of switch loops earlier if a slow mon is involved)