gym rework changes batch 1
Reworked the Gym Leader Events in RPGMaker to correctly link to the new teams, and load you into Normal/Cursed based on Amulet Status instead of dialogue choices.
Changed the event triggered on Perfecting a Gym Leader giving you the Key to the Avatar, to instead be triggered on completing the Cursed fight.
Changed the first Imogene event to be unskippable.