Desert + Estuary trainers
Desert trainer now a hard sun team with hard setting weather, to contrast with Nyx whos post-rock climb but on the same route. Sudo is the only setter so if its eliminated the weather can be removed and kneecap the rest of the fight, but Prankster makes it hard to stop the 1st set.

also timer balls nearby are 6 instead of 3

ReplaceMe2 goes in the Estuary, contrasting with Yvvone with some bulky hazard stackers, with some tricks to force you to swap and take hazard chip