Hidden Groudon/Kyogre route changes
Made the hiker on the Volcanic Shore required
Added a Clerk (Land development) in the Boiling Cavern, replacing the Avatar of Octillery.
Added more forced encounter tiles on the path to Groudon.
Made the Firebreather in the Underground River required
Add a Blackbelt trainer in the Abyssal Cavern, replacing the Avatar of Lanturn
Made the Avatar of Slowking optional, and had it drop the Water Token the Lanturn used to drop
Add more forced encounter tiles on the path to Kyogre.