Miltank rework pt.2
Reworked Miltank's stats again to make it more of a dedicated Fast Wall (with high HP for Wish purposes)
Stability replaces Breakthrough on it

Added new ability "Refreshing": "This Pokémon's moves heal for 25% more."
Added code support for the healing helper methods being able to care about the user of the move that caused the healing (when relevant). Draining type effects are considered to be "from the user's move" but effects placed by moves (e.g. Aqua Ring) are not. Shell Bell, despite being not technically draining (it doesn't synergize with Big Root, Gorging, etc.) is still counted as "healing from moves" for this purpose.
For Wish, I refactored Foretold Moves and Wish to both use a new helper method of the Battle class called getUserForDelayedMove. That way, Wish healing can benefit from Refreshing even when the user has swapped out.
Also refactored the move codes for Wish and Arc of Hope so that Arc of Hope now inherits from Wish.